Added state/apply/delete impl.
Allows to delete packages from the cluster by name.

To delete packages gracefully (avoiding finalizer deadlocks)
the objects (cluster state returned by kubectl) are grouped into
crd resources, namespaced, global and crds and deleted in this order
waiting for each group to be deleted before deleting the next group.
However this order definition is a first draft and may need to evolve
(as in kustomize).